chainlit_app/requirements.txt
python-dotenv==1.0.1
chainlit==2.0.3
requests==2.32.3
